requirements.txt
annotated-types==0.7.0
anyio==4.12.1
astrapy==2.1.0
certifi==2026.1.4
charset-normalizer==3.4.4
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.7
langchain-astradb==1.0.0
langchain-classic==1.0.1
langchain-core==1.2.7
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langchain-community>=0.2.0
langchainhub==0.1.21
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.7
numpy==2.4.2
openai==2.16.0
orjson==3.11.6
ormsgpack==1.12.2
packaging==24.2
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.46
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
tqdm==4.67.2
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid6==2025.0.1
uuid_utils==0.14.0
xxhash==3.6.0
zstandard==0.25.0
